requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
emails==0.6
jinja2==3.1.2
python-dotenv==1.0.0
pydantic==2.5.1
pydantic-settings==2.1.0
email-validator==2.1.0
